fix: silence MSVC C4244 double-to-float narrowing in Peak1D emplace_back (#8921)
Peak1D::IntensityType is float, but many emplace_back calls passed
double literals (1.0) or double variables as the intensity argument,
triggering C4244 "possible loss of data" warnings on MSVC.

- Use 1.0f literals where intensity is a constant
- Add static_cast<float>() where intensity comes from double variables
  or expressions (TheoreticalSpectrumGenerator intensity parameters
  and pre_int_ members)

+++ src/openms/source/CHEMISTRY/TheoreticalSpectrumGenerator.cpp
@@ -322,7 +322,7 @@ namespace OpenMS
         ion_names.emplace_back("iP+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(70.0656, 1.0); // emplace_back(MZ, intensity)
+      spectrum.emplace_back(70.0656, 1.0f); // emplace_back(MZ, intensity)
     }
 
     // Cysteine (C2H6NS)
@@ -333,7 +333,7 @@ namespace OpenMS
         ion_names.emplace_back("iC+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(76.0221, 1.0);
+      spectrum.emplace_back(76.0221, 1.0f);
     }
 
     // Iso/Leucin immonium ion (same mass for immonium ion)
@@ -344,7 +344,7 @@ namespace OpenMS
         ion_names.emplace_back("iL/I+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(86.09698, 1.0);
+      spectrum.emplace_back(86.09698, 1.0f);
     }
 
     // Histidin immonium ion (C5H8N3)
@@ -355,7 +355,7 @@ namespace OpenMS
         ion_names.emplace_back("iH+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(110.0718, 1.0);
+      spectrum.emplace_back(110.0718, 1.0f);
     }
 
     // Phenylalanin immonium ion (C8H10N)
@@ -366,7 +366,7 @@ namespace OpenMS
         ion_names.emplace_back("iF+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(120.0813, 1.0);
+      spectrum.emplace_back(120.0813, 1.0f);
     }
 
     // Tyrosine immonium ion (C8H10NO)
@@ -377,7 +377,7 @@ namespace OpenMS
         ion_names.emplace_back("iY+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(136.0762, 1.0);
+      spectrum.emplace_back(136.0762, 1.0f);
     }
 
     // Tryptophan immonium ion
@@ -388,7 +388,7 @@ namespace OpenMS
         ion_names.emplace_back("iW+");
         charges.push_back(1);
       }
-      spectrum.emplace_back(159.0922, 1.0);
+      spectrum.emplace_back(159.0922, 1.0f);
     }
   }
 
@@ -447,7 +447,7 @@ namespace OpenMS
     // TODO why do you need a separate set for the losses? Just use the keys from the formula_str_cache?
     for (const auto& formula : f_losses)
     {
-      spectrum.emplace_back((mz - formula.getMonoWeight()) / (double)charge, intensity);
+      spectrum.emplace_back((mz - formula.getMonoWeight()) / (double)charge, static_cast<float>(intensity));
 
       if (add_metainfo)
       {
@@ -535,7 +535,7 @@ namespace OpenMS
             ion_names.push_back(ion_name);
             charges.push_back(charge);
           }
-          spectrum.emplace_back(iso.getMZ() / (double)charge, intensity * rel_loss_intensity_ * iso.getIntensity());
+          spectrum.emplace_back(iso.getMZ() / (double)charge, static_cast<float>(intensity * rel_loss_intensity_ * iso.getIntensity()));
         }
       }
       else
@@ -545,7 +545,7 @@ namespace OpenMS
           ion_names.push_back(ion_name);
           charges.push_back(charge);
         }
-        spectrum.emplace_back(loss_pos / (double)charge, intensity * rel_loss_intensity_);
+        spectrum.emplace_back(loss_pos / (double)charge, static_cast<float>(intensity * rel_loss_intensity_));
       }
     }
   }
@@ -639,7 +639,7 @@ namespace OpenMS
         }
         pos = (pos + ion_offset) / charge;
 
-        spectrum.emplace_back(pos, intensity);
+        spectrum.emplace_back(pos, static_cast<float>(intensity));
         if (add_metainfo_)
         {
           if (res_type == Residue::AIon)
@@ -810,7 +810,7 @@ namespace OpenMS
           }
           pos = (pos + ion_offset) / charge;
 
-          spectrum.emplace_back(pos, intensity);
+          spectrum.emplace_back(pos, static_cast<float>(intensity));
           if (add_metainfo_)
           {
             ion_names.emplace_back(ion_name_str);
@@ -910,7 +910,7 @@ namespace OpenMS
           }
           pos = (pos + ion_offset) / charge;
 
-          spectrum.emplace_back(pos, intensity);
+          spectrum.emplace_back(pos, static_cast<float>(intensity));
           if (add_metainfo_)
           {
             ion_names.emplace_back(ion_name_str);
@@ -1025,7 +1025,7 @@ namespace OpenMS
           ion_names.push_back(ion_name);
           charges.push_back(charge);
         }
-        spectrum.emplace_back(it->getMZ() / (double)charge, pre_int_ * it->getIntensity());
+        spectrum.emplace_back(it->getMZ() / (double)charge, static_cast<float>(pre_int_ * it->getIntensity()));
       }
     }
     else
@@ -1035,7 +1035,7 @@ namespace OpenMS
         ion_names.push_back(ion_name);
         charges.push_back(charge);
       }
-      spectrum.emplace_back(mono_pos / (double)charge, pre_int_);
+      spectrum.emplace_back(mono_pos / (double)charge, static_cast<float>(pre_int_));
     }
     // loss peaks of the precursor
 
@@ -1074,7 +1074,7 @@ namespace OpenMS
           ion_names.push_back(ion_name);
           charges.push_back(charge);
         }
-        spectrum.emplace_back(it->getMZ() / charge, pre_int_H2O_ * it->getIntensity());
+        spectrum.emplace_back(it->getMZ() / charge, static_cast<float>(pre_int_H2O_ * it->getIntensity()));
       }
     }
     else
@@ -1093,7 +1093,7 @@ namespace OpenMS
         ion_names.push_back(ion_name);
         charges.push_back(charge);
       }
-      spectrum.emplace_back(mono_pos / (double)charge, pre_int_H2O_);
+      spectrum.emplace_back(mono_pos / (double)charge, static_cast<float>(pre_int_H2O_));
     }
 
     //loss of ammonia
@@ -1133,7 +1133,7 @@ namespace OpenMS
           ion_names.push_back(ion_name);
           charges.push_back(charge);
         }
-        spectrum.emplace_back(it->getMZ() / (double)charge, pre_int_NH3_ * it->getIntensity());
+        spectrum.emplace_back(it->getMZ() / (double)charge, static_cast<float>(pre_int_NH3_ * it->getIntensity()));
       }
     }
     else
@@ -1152,7 +1152,7 @@ namespace OpenMS
         ion_names.push_back(ion_name);
         charges.push_back(charge);
       }
-      spectrum.emplace_back(mono_pos / (double)charge, pre_int_NH3_);
+      spectrum.emplace_back(mono_pos / (double)charge, static_cast<float>(pre_int_NH3_));
     }
   }
